jobs_platform/requirements.txt
django >= 5
psycopg[binary]
djangorestframework